chore: Move CSS properties code to own file
